+++ storage/backend.go
@@ -988,7 +988,6 @@ func (b *StorageBackend) ConstructExternal(ctx context.Context) *BackendExternal
 type PersistentStorageBackendConfig struct {
 	OntapConfig             *drivers.OntapStorageDriverConfig     `json:"ontap_config,omitempty"`
 	SolidfireConfig         *drivers.SolidfireStorageDriverConfig `json:"solidfire_config,omitempty"`
-	EseriesConfig           *drivers.ESeriesStorageDriverConfig   `json:"eseries_config,omitempty"`
 	AWSConfig               *drivers.AWSNFSStorageDriverConfig    `json:"aws_config,omitempty"`
 	AzureConfig             *drivers.AzureNFSStorageDriverConfig  `json:"azure_config,omitempty"`
 	GCPConfig               *drivers.GCPNFSStorageDriverConfig    `json:"gcp_config,omitempty"`
@@ -1005,8 +1004,6 @@ func (psbc *PersistentStorageBackendConfig) GetDriverConfig() (drivers.DriverCon
 		driverConfig = psbc.OntapConfig
 	case psbc.SolidfireConfig != nil:
 		driverConfig = psbc.SolidfireConfig
-	case psbc.EseriesConfig != nil:
-		driverConfig = psbc.EseriesConfig
 	case psbc.AWSConfig != nil:
 		driverConfig = psbc.AWSConfig
 	case psbc.AzureConfig != nil:
@@ -1063,8 +1060,6 @@ func (p *BackendPersistent) MarshalConfig() (string, error) {
 		bytes, err = json.Marshal(p.Config.OntapConfig)
 	case p.Config.SolidfireConfig != nil:
 		bytes, err = json.Marshal(p.Config.SolidfireConfig)
-	case p.Config.EseriesConfig != nil:
-		bytes, err = json.Marshal(p.Config.EseriesConfig)
 	case p.Config.AWSConfig != nil:
 		bytes, err = json.Marshal(p.Config.AWSConfig)
 	case p.Config.AzureConfig != nil: